chore(workflows): Labeler workflow 추가 (#385)
* chore: add github labeler configuration

This commit introduces a `.github/labeler.yml` file
to automate issue and pull request labeling based on
file paths.

This configuration defines rules for automatically
applying labels like "package: api", "package: icon",
"eslint", "tsconfig", "documentation", "build",
"workflows", "app: spectator", and "app: manager" to
issues and pull requests based on the files that have
been changed.

This enhances workflow efficiency by automatically
categorizing and organizing issues and pull requests,
making it easier to identify areas of the codebase
affected by specific changes.

* chore: upgrade actions/checkout to v4

This commit upgrades the actions/checkout action to v4 in the
sync_spectator workflow. This upgrade is necessary to stay up-to-date
with the latest security patches and features provided by the action.
The change addresses potential vulnerabilities and ensures compatibility
with newer versions of GitHub Actions.

There are no breaking changes introduced by this upgrade.

* chore: enforce newline at end of file for labeler workflow

Ensures the .github/workflows/labeler.yml file has a newline character at the end of the file. This change is necessary to adhere to standard file formatting conventions and prevent potential issues with tools or scripts that rely on this convention. No breaking changes are introduced.